REFACTOR: namespaces, class consolidation
