+++ Assets/Prefabs/PlayerBelongings.prefab
@@ -2731,8 +2731,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 47248e9f01165794980cd170970e3816, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  team: {fileID: 0}
   health: {fileID: 11400000, guid: 95706906eda1dbf42bf0b2b83411d86f, type: 2}
   maxHealth: {fileID: 11400000, guid: 4b9ff31e6512a264fafb5e372d80ff92, type: 2}
+  armor: {fileID: 0}
   speed: {fileID: 11400000, guid: da106d5806969c14693d9c523cd6f87d, type: 2}
   sneakMultiplier: 0.7
   runMultiplier: 2
@@ -2740,6 +2742,7 @@ MonoBehaviour:
   weight: {fileID: 11400000, guid: 614a093e4ff687f47b74804bf6ef4bea, type: 2}
   maxWeight: {fileID: 11400000, guid: 1b6144f0b378a564184f78a60f5a177e, type: 2}
   slotCapacity: -1
+  damageCounter: {fileID: 0}
 --- !u!114 &8379673560634840537
 MonoBehaviour:
   m_ObjectHideFlags: 0
