+++ Assets/Scripts/Character/CharacterProperties.cs.meta
