+++ Assets/Scripts/Character/NPC.meta
