+++ Assets/Scripts/Character/NPC/NPCProperties.cs.meta
