+++ Assets/Scripts/Character/Player/PlayerProperties.cs
@@ -1,0 +1,8 @@
+﻿using Character.NPC;
+
+namespace Character.Player
+{
+    public class PlayerProperties : CharacterProperties
+    {
+    }
+}
