+++ Assets/Scripts/Character/Player/PlayerProperties.cs.meta
