--- Assets/Scripts/Player/PlayerProperties.cs
@@ -1,6 +1,0 @@
-﻿using NPC;
-
-public class PlayerProperties : CharacterProperties
-{
-    
-}
