+++ Assets/Scripts/Stat/AttributeType.cs
@@ -4,52 +4,42 @@ using UnityEngine;
 namespace Stat
 {
     /// <summary>
-    /// Attribute Type.
+    ///     Attribute Type.
     /// </summary>
     [CreateAssetMenu(menuName = "Stat/AttributeType")]
     public class AttributeType : ScriptableObject, IEquatable<AttributeType>
     {
-        public string Type => type;
-
         /// <summary>
-        /// Identifier.
+        ///     Identifier.
         /// </summary>
         [Tooltip("Identifier String")] [SerializeField]
         private string type;
 
+        public string Type => type;
+
+        public bool Equals(AttributeType other)
+        {
+            if (ReferenceEquals(null, other)) return false;
+
+            if (ReferenceEquals(this, other)) return true;
+
+            return base.Equals(other) && type == other.type;
+        }
+
         public static AttributeType Create(string type)
         {
+            if (type == null) throw new ArgumentException("type must not be null");
+
             var attributeType = CreateInstance<AttributeType>();
             attributeType.type = type;
             return attributeType;
         }
 
-        public bool Equals(AttributeType other)
-        {
-            if (ReferenceEquals(null, other))
-            {
-                return false;
-            }
-
-            if (ReferenceEquals(this, other))
-            {
-                return true;
-            }
-
-            return base.Equals(other) && type == other.type;
-        }
-
         public override bool Equals(object obj)
         {
-            if (ReferenceEquals(null, obj))
-            {
-                return false;
-            }
+            if (ReferenceEquals(null, obj)) return false;
 
-            if (ReferenceEquals(this, obj))
-            {
-                return true;
-            }
+            if (ReferenceEquals(this, obj)) return true;
 
             return Equals((AttributeType) obj);
         }
