+++ Assets/Scripts/Status/PoisonEffect.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using Character.Player;
+using UnityEngine;
 
 namespace Status
 {
