--- Assets/Scripts/Util/RangedFloatVariableClamp.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: a53cc67d9cc94228bd81592d26937033
-timeCreated: 1587076660
